pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptinjector"
version = "0.1.0"
description = "Security testing tool for prompt injection vulnerabilities in custom GPTs and Gems"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
keywords = ["security", "prompt-injection", "llm", "gpt", "testing"]
dependencies = [
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
promptinjector = "promptinjector.cli:main"

[project.urls]
Homepage = "https://github.com/billford/promptInjector"
Issues = "https://github.com/billford/promptInjector/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
promptinjector = ["injections/payloads/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
